Memory usage now in size_t as it should be

+++ headers/RadarKit/RKRadar.h
@@ -49,7 +49,7 @@ typedef struct RKRadar {
     RKRadarState               state;
     bool                       active;
     //
-    unsigned int               memoryUsage;
+    size_t                     memoryUsage;
     //
     //
     //
